pyproject.toml
[tool.poetry]
name = "ProgImage"
version = "1.0.0"
description = "Image storage and manipulation engine"
authors = ["Mate Pager <mate.pager@gmail.com>"]
readme = "README.md"
include = ["README.md"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
gunicorn = "^20.0.4"
boto3 = "^1.16.18"
Pillow = "^8.0.1"
certifi = "^2020.11.8"
requests = "^2.25.0"
requests-toolbelt = "^0.9.1"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
pytest-flask = "^1.1.0"
moto = "^1.3.16"
requests-mock = "^1.8.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
